prepare for first gyro tests
