+++ src/hardware_interfaces/test/test_gyro.py
@@ -1,0 +1,10 @@
+import hardware_interfaces.gyro
+
+def test_pony():
+    assert 11==2
+
+def test_tests():
+    pass
+
+
+
